level1_example.py passing but crash on report generation
