+++ aviary/api.py
@@ -96,7 +96,7 @@ from aviary.subsystems.propulsion.utils import build_engine_deck
 ###################
 
 # Miscellaneous
-from aviary.interface.methods_for_level2 import PreMissionGroup, PostMissionGroup
+# from aviary.interface.methods_for_level2 import PreMissionGroup, PostMissionGroup
 from aviary.subsystems.premission import CorePreMission
 from aviary.subsystems.subsystem_builder_base import SubsystemBuilderBase
 from aviary.utils.preprocessors import (
